# LoongArch64: add lsx support