@lemire 2025-07-10
@Xiao-Tao 

Could we get CI tests to work?
https://github.com/simdutf/simdutf/blob/master/.github/workflows/loongarch64-gcc-14.2.yml